+++ CHANGELOG.md
@@ -50,6 +50,7 @@ This allows using `distortos::Mutex` with all mutex helpers from
 - Added unit tests of *STM32's* *DMAv1*, *DMAv2*, *SPIv1* and *SPIv2* drivers.
 - Added unit tests of all `estd::ContiguousRange` constructor overloads.
 - Added `estd/EnumClassFlags.hpp` which provides templated bitwise operators for enum class flags.
+- Added `estd/log2u.hpp` with `log2()`-like `constexpr` function for `unsigned int`.
 
 ### Changed